Add test for Customer attributes and convert to TS
